Create app primary endpoint ("/checkAccess")
This endpoint is the main way to access the API.
At the moment, it receives a GET request and then simply prints it to the console.

While this is obviously an initial commit for this funcionality,
it ignores some problems as outlined in the assignment description.
- Information is sent to this endpoint via URI encoding, rather than JSON.
- An extension to that problem is that information is visible in the address bar.

Ideally, I switch over to a POST request and send the data as a JSON block later on.