Create fileManager and get data from text files
My solution to persistent data for the list of enrolled sellers,
pre-approved categories, and the minimum price, was simply to use
text files. I created a fileManager which, as of this commit,
only reads files.

This solution was chosen because it was fast to implement,
and functional enough!

This is not a secure solution, this is not a production quality solution...
It is just a quick fix!